stuff
- removed shield method from encounter because its no longer used
- removed time library since delay is implemented with tk
- removed string library since tags don't rely on random names anymore
- added music for glados fight and aperture lap
- moved encountertest out of content
